Fix addMetaDataLine so sequence dictionary not duplicated
There was an issue in addMetaDataLine where it was calling
loadMetaDataMaps after adding a new header line, which repopulated
all metadata again.  This was mostly not a problem since the
majority of the data structures were maps, so the keys/values were
just replaced, however contigs were stored in a list, so the list
effectively got duplicated.  Instead the functionality inside
loadMetaDataMaps got moved into a function to handle one line
at a time, which now gets called by addMetaDataLine for the new
line, and in loadMetaDataMaps inside a for-loop for each
metadata entry.  Only the constructor makes use of loadMetaDataMaps.
Unit test added to VCFHeaderUnitTest.java to ensure the exception
is no longer thrown.

Resolves: 214