Modify IndexFactory.createTabixIndex() for bgzip input
Modify the internal FeatureIterator.initStream method inside
IndexFactory to accept both block-compressed and non-compressed
input files. This method leaves the constructor as is and does
the work without extra work for the user.  The method will
check that this is proper bgzip data, and throw exception if
not.  The unit test checks that the same result comes from calling
this method with a VCF and the same bgzipped VCF.

Resolves: 184